Optional verificat amb ".isPresent()". A client el numero de tarjeta de crèdit ara és Long per a poder introduïr la quantitat de xifres necessitades.

+++ src/main/java/Projecte2/DriftoCar/entity/MySQL/Client.java
@@ -57,7 +57,7 @@ public class Client {
     private LocalDate dniCaducitat;
     
     @Column(name = "num_tarjeta_credit", unique = true)
-    private Integer numTarjetaCredit;
+    private Long numTarjetaCredit;
     
     private String adreca;
